requirements.txt
opencv-python~=4.7.0.68
pandas~=1.5.2
streamlit~=1.18.1
numpy~=1.23.0
torch~=1.13.0
omegaconf~=2.2.3
tqdm~=4.64.1

Pillow~=9.4.0
